+++ CryptokiMan/encrypt_thread.cpp
@@ -40,7 +40,7 @@ void EncryptThread::run()
     int nPartSize = manApplet->settingsMgr()->fileReadSize();
     qint64 nReadSize = 0;
     int nLeft = 0;
-    int nOffset = 0;
+    qint64 nOffset = 0;
 
     BIN binPart = {0,0};
     BIN binDst = {0,0};
